[vcpkg] Rewrite process spawning on windows to always use CreateProcess, enabling better environment handling
